Split AI provider DTOs into one file per type
Each record now lives in its own .cs file in both the API and Components
layers. Deleted the combined AiProviderDtos.cs file.
