+++ code/FinanceManager.Api/Controllers/Admin/AiConfigurationResponse.cs
@@ -1,0 +1,5 @@
+namespace FinanceManager.Api.Controllers.Admin;
+
+public sealed record AiConfigurationResponse(
+    IReadOnlyList<AiProviderDto> Providers,
+    IReadOnlyList<AiFallbackEntryDto> FallbackEntries);
