+++ code/FinanceManager.Api/Controllers/Admin/AiFallbackEntryDto.cs
@@ -1,0 +1,3 @@
+namespace FinanceManager.Api.Controllers.Admin;
+
+public sealed record AiFallbackEntryDto(string ProviderName, string Model, int Order);
