+++ code/FinanceManager.Api/Controllers/Admin/AiProviderDto.cs
@@ -1,0 +1,8 @@
+namespace FinanceManager.Api.Controllers.Admin;
+
+public sealed record AiProviderDto(
+    string ProviderName,
+    string BaseUrl,
+    bool HasApiKey,
+    int RequestTimeoutSeconds,
+    bool IsEnabled);
