+++ code/FinanceManager.Api/Controllers/Admin/UpdateFallbackRequest.cs
@@ -1,0 +1,4 @@
+namespace FinanceManager.Api.Controllers.Admin;
+
+public sealed record UpdateFallbackRequest(
+    IReadOnlyList<AiFallbackEntryDto> Entries);
