+++ code/FinanceManager.Api/Controllers/Admin/UpdateProviderRequest.cs
@@ -1,0 +1,7 @@
+namespace FinanceManager.Api.Controllers.Admin;
+
+public sealed record UpdateProviderRequest(
+    string BaseUrl,
+    string? ApiKey,
+    int RequestTimeoutSeconds,
+    bool IsEnabled);
