+++ code/FinanceManager.Components/HttpClients/AiConfigurationResponse.cs
@@ -1,0 +1,5 @@
+namespace FinanceManager.Components.HttpClients;
+
+public sealed record AiConfigurationResponse(
+    List<AiProviderDto> Providers,
+    List<AiFallbackEntryDto> FallbackEntries);
