+++ code/FinanceManager.Components/HttpClients/AiFallbackEntryDto.cs
@@ -1,0 +1,3 @@
+namespace FinanceManager.Components.HttpClients;
+
+public sealed record AiFallbackEntryDto(string ProviderName, string Model, int Order);
