+++ code/FinanceManager.Components/HttpClients/AiProviderDto.cs
@@ -1,0 +1,8 @@
+namespace FinanceManager.Components.HttpClients;
+
+public sealed record AiProviderDto(
+    string ProviderName,
+    string BaseUrl,
+    bool HasApiKey,
+    int RequestTimeoutSeconds,
+    bool IsEnabled);
