+++ code/FinanceManager.Components/HttpClients/UpdateFallbackRequest.cs
@@ -1,0 +1,3 @@
+namespace FinanceManager.Components.HttpClients;
+
+public sealed record UpdateFallbackRequest(List<AiFallbackEntryDto> Entries);
