+++ code/FinanceManager.Components/HttpClients/UpdateProviderRequest.cs
@@ -1,0 +1,7 @@
+namespace FinanceManager.Components.HttpClients;
+
+public sealed record UpdateProviderRequest(
+    string BaseUrl,
+    string? ApiKey,
+    int RequestTimeoutSeconds,
+    bool IsEnabled);
